Initial commit
Initial steps towards a food tracking website. This commit adds a zig
webserver w/ an sqlite database and a lil' bit of web visualization of
DB. We allow entry of "properties" and "ingredients". Properties are
things you would see on a food label, ingredients are things you would
buy at the store

Eventually these will be tied into meals etc, but we need to start
somewhere