Add meals/dishes
Add concept of meals and dishes for tracking daily food intake. The idea
here is that dishes are categories of food (e.g. burger, or egg and
bread), and meals are a collection of dishes

Dishes do NOT have ingredients. This is because I do not measure
ingredients, or even necessarily use the same ingredients for the same
type of food. I figure I care more about "i ate burger" then "i ate this
specific burger with this combo of ingredients that i named". I know I
will never do the second one, so the first is more reasonable

Implementation details:
* update() function added to RemoteItemList abstraction in JS land, this
  is to allow us to handle back button presses on the calendar view
* Added APIs for meals/dishes/meal dishes/meal dish ingredients

Bonus changes:
* Improve error messages in db
* Fix port collision in integration test
* Improve error messages on sqlite statement execution failures