Add helper class to manage remote array of stuff
A common pattern (present, and being used more in future) is to take a
list of resources on the server, present them in some way, and add new
elements. On addition we want to display those items as well.

Add a helper class to help manage this list of things and call a
callback when new items are added. This heavily simplifies our
ingredient/property list pages