Add delete meal button
Delete a meal. On deletion we propagate deletions all the way down. No
confirmation, for now it seems like if we delete a meal it's easy to
re-add. If we end up regretting this decision in the future we can
always add a confirmation dialog later

Sqlite has an ON CASCADE feature that only works when foregin key
constraints are explicitly enabled. Set up and add a rare DB unit test
to ensure this works as expected. Make sure we run the tests in the
linter script as well