Support decimal entry in property values
When testing, we found that it is hard to predict a property unit that
will never have decimals. Allow decimal entry

API represents property values as fixed point numbers up to 3 decimals.
This ensures that we take _exactly_ what was written in the form instead
of hoping that a float rounds correctly. We then cast our fixed point
repr into the DB and back.

Summaries also use the FixedPointNumber abstraction to avoid giving the
webapp xxx.yyyyyyyyyyyyyy