Slight tweak to sphdelete-button css
We really struggled here to get the layout correct. I believe we had the
following sequence

* water.css is setting input display type to block. This results in the
  hitbox of the surrounding element extending all the way to the right
  edge of the screen. Because we set the click events on the parent
  element, clicking anywhere on the line would trigger a deletion
* We set the input type to inline-block to force the surrounding element
  to treat it like an inline element and not extend to the right
* inline elements are rendered like text, so the surrounding element pads
  its box to ensure that the line height is respected
* Eventually switch the surrounding div to contents to avoid that

After doing a bunch of reading, I believe this is a more reasonable
solution. Set size constraints on the root element and attempt to fill
that space with the input element. This allows external users to
interact with the sphdelete-button like it's a normal element.