Search for ingredients in meal entry
Current ingredient entry is quite painful. Ingredients are not sorted in
a way that is useful and there are a lot of ingredients. Adding to meal
view should be quick and easy. Replace ingredient dropdown with a search
box

Delete button used to be next to ingredients dropdown, this no longer
makes sense. Move delete button to the left of the title with a fancy
lil' x

Implementation details:
* dish_buttons group is no longer needed
* Some of the sphearch widget construction needed to be shuffled around
  * Attributes are only available on connected
  * We try to set search results before connecting to the dom
  * Move inner div construction to sphearch constructor
* Modify Sphearch widget to hide search results by default and only show
  on click
  * Showing all ingredients all the time made it hard to scroll the meal
    view
* The rest is hopefully self explanatory :D